+++ src/burst2safe/burst2safe.py
@@ -50,7 +50,7 @@ def find_group(orbit: int, footprint: Polygon, polarizations: Iterable[str]) -> 
         if not single_pol:
             # TODO: add Vertex link to error message?
             raise ValueError(
-                f'No results found for orbit {orbit}, footprint {footprint}, and polarization {pol}.'
+                f'No results found for orbit {orbit}, footprint {footprint}, and polarization {pol}. '
                 'Bursts may not populated yet for this group. Check Vertex to confirm.'
             )
         results.extend(list(single_pol))
@@ -153,7 +153,7 @@ def main() -> None:
     parser = ArgumentParser(description=DESCRIPTION)
     parser.add_argument('granules', nargs='*', help='A list of bursts to convert to SAFE')
     parser.add_argument('--orbit', type=int, help='The absolute orbit number of the bursts')
-    parser.add_argument('--bbox', type=float, nargs=4, help='Bounding box of the bursts (S W N E in lat/lon)')
+    parser.add_argument('--bbox', type=float, nargs=4, help='Bounding box of the bursts (W S E N in lat/lon)')
     parser.add_argument('--pols', type=str, nargs='+', help='The polarizations of the bursts (i.e., VV VH)')
     parser.add_argument('--keep-files', action='store_true', default=False, help='Keep the intermediate files')
     args = parser.parse_args()